move tool output base components to a dedicated file
